Refactoring of merging ai clients into one
